vault: avoid continual renewal of invalid token
A series of errors may happen when a token is invalidated while the
Vault client is waiting to renew it. The token may have been invalidated
for several reasons, such as the alloc finished running and it's now
terminal or the token may have been change directly on Vault
out-of-band.

Most of the errors are caused by retries that will never succeed until
Vault fully removes the token from its state.

This commit prevents the retries by making the error `invalid lease ID`
a fatal error.

In earlier versions of Vault, this case was covered by the error `lease
not found or lease is not renewable`, which is already considered to be
a fatal error by Nomad:

https://github.com/hashicorp/vault/blob/2d0cde4ccc0323591d9414342cb15f5cb70271d7/vault/expiration.go#L636-L639

But https://github.com/hashicorp/vault/pull/5346 introduced an earlier
`nil` check that generates a different error message:

https://github.com/hashicorp/vault/blob/750ab337eaa0b049d9cf1535c00e860129e5e9a0/vault/expiration.go#L1362-L1364

Both errors happen for the same reason (`le == nil`) and so should be
considered fatal on renewal.

There is also a possible race condition in `vaultclient.run()` loop in
the following `select` statement:

https://github.com/hashicorp/nomad/blob/78f0c6b2a910c25522ac88ff63343bb371d72bff/client/vaultclient/vaultclient.go#L557-L564

When the `vault_hook` stops, it cancels the token renewal, which
triggers the `updateCh` to re-run the renewal loop and re-fetch the next
token to renew.

But it may happen that `renewalCh` and `updateCh` are both active, such
as when a task exits very closely to token renewal time. In this
scenario, the `case` selected is random and, if the renewal case is
selected, the token is incorrectly re-added to the heap, causing
unnecessary renewals of a token that is already expired.

To prevent this, this commits adds a check to exit if the token is no
longer tracked by the heap, suggesting that a call to `StopRenewToken()`
happened while the `vaultclient` was waiting in the `select`.

Since both `StopRenewToken()` and `renew()` hold a lock to the heap
throughout their execution, checking `isTracked()` once is enough as the
heap would not be modified until either one returns.

Given how time sensitive this issue is, it was not possible to create a
reliable reproduction nor tests. The fixes were applied after analysing
the code paths in Nomad and Vault that could generate the reported
behaviour.